+++ Sources/CodexBar/SettingsStoreState.swift
@@ -30,7 +30,6 @@ struct SettingsDefaultsState {
     var menuBarDisplayModeRaw: String?
     var kiroMenuBarDisplayModeRaw: String?
     var historicalTrackingEnabled: Bool
-    var historicalPaceMinimumWeeks: Int
     var multiAccountMenuLayoutRaw: String
     var menuBarMetricPreferencesRaw: [String: String]
     var copilotBudgetExtrasEnabled: Bool